requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
python-dotenv
openai
httpx
jinja2
xgboost
pytest
scikit-learn